Добавляет изменения в js
